refactor: fixed styles for sidebars and normalised select component
